% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams_finder.R
\name{gams_finder}
\alias{gams_finder}
\title{GAMS finder for windows machines}
\usage{
gams_finder()
}
\value{
GAMS installation location
}
\description{
GAMS finder for windows machines
}
\examples{
\dontrun{
gams_finder()
}
}
