% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_KIRAN.R
\name{install_KIRAN}
\alias{install_KIRAN}
\title{Install dependency packages needed needed for KIRAN suite of packages}
\usage{
install_KIRAN()
}
\description{
Install dependency packages needed needed for KIRAN suite of packages
}
\examples{
\dontrun{
install_KIRAN()
}
}
