% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools_finder.R
\name{rtools_finder}
\alias{rtools_finder}
\title{Rtools finder for windows machines}
\usage{
rtools_finder()
}
\value{
Message related to Rtools check
}
\description{
Rtools finder for windows machines
}
\examples{
\dontrun{
rtools_finder()
}
}
