% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanGAMS.R
\name{scanGAMS}
\alias{scanGAMS}
\title{GAMS finder for windows machines}
\usage{
scanGAMS()
}
\value{
GAMS installation location
}
\description{
GAMS finder for windows machines
}
\examples{
\dontrun{
scanGAMS()
}
}
