% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanRtools.R
\name{scanRtools}
\alias{scanRtools}
\title{Rtools finder for windows machines}
\usage{
scanRtools()
}
\value{
Message related to Rtools check
}
\description{
Rtools finder for windows machines
}
\examples{
\dontrun{
scanRtools()
}
}
