% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKIRAN.R
\name{setKiran}
\alias{setKiran}
\title{Install dependency packages needed needed for KIRAN suite of packages}
\usage{
setKiran()
}
\description{
Install dependency packages needed needed for KIRAN suite of packages
}
\examples{
\dontrun{
setKiran()
}
}
