% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLibs.R
\name{setLibs}
\alias{setLibs}
\title{Set user specific R library for package installation}
\usage{
setLibs(export = FALSE)
}
\arguments{
\item{export}{If R_LIBS_USER setting should be returned}
}
\description{
Set user specific R library for package installation
}
\examples{
\dontrun{
setLibs()
}
}
