% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRepos.R
\name{setRepos}
\alias{setRepos}
\title{Setting Repositories}
\usage{
setRepos()
}
\value{
Sets IFPRI based R repositories related to IMPACT model
}
\description{
Setting Repositories
}
\examples{
\dontrun{
setRepos()
}
}
